+++ lib/prompt/command.ex
@@ -32,8 +32,8 @@ defmodule Prompt.Command do
 
     @impl true
     def init(_argv) do
-      # parse list of args to map or struct
-      %{list: true, help: false, directory: "path/to/dir"}
+      # parse list of args to a struct if desired
+      %SomeStruct{list: true, help: false, directory: "path/to/dir"}
     end
 
     @impl true
@@ -42,27 +42,6 @@ defmodule Prompt.Command do
       display(File.ls!(dir))
     end
 
-  end
-  ```
-
-  Typically one will use the `OptionParser.parse/1` function to parse
-  the command
-
-  ```
-  defp parse(argv) do
-   argv
-   |> OptionParser.parse(
-    strict: [help: :boolean, directory: :string, list: :boolean],
-    aliases: [h: :help, d: :directory]
-   )
-   |> _parse()
-  end
-
-  defp _parse({opts, _, _}) do
-   help = Keyword.get(opts, :help, false)
-   dir = Keyword.get(opts, :directory, "./")
-   list = Keyword.get(opts, :length, true)
-   %{help: help, directory: dir, list: list}
   end
   ```
 
@@ -81,7 +60,7 @@ defmodule Prompt.Command do
   Takes the options passed in via the command line and
   tramforms them into a struct that the process command can handle
   """
-  @callback init(list(String.t())) :: term
+  @callback init(map()) :: term
 
   @doc """
   Prints the help available for this command
@@ -99,6 +78,7 @@ defmodule Prompt.Command do
       import Prompt
 
       @doc false
+      @impl Prompt.Command
       def help() do
         help =
           case Code.fetch_docs(__MODULE__) do
@@ -111,32 +91,30 @@ defmodule Prompt.Command do
         display(help)
       end
 
-      @before_compile Prompt.Command
-
       defoverridable help: 0
+
+      @doc false
+      @impl Prompt.Command
+      def init(init_arg) do
+        init_arg
+      end
+
+      defoverridable init: 1
+
+      @before_compile Prompt.Command
     end
   end
 
   defmacro __before_compile__(env) do
-    unless Module.defines?(env.module, {:init, 1}) do
+    unless Module.defines?(env.module, {:process, 1}) do
       message = """
-      function init/1 required by behaviour Prompt.Command is not implemented \
+      function process/1 required by behaviour Prompt.Command is not implemented \
       (in module #{inspect(env.module)}).
 
-      You'll need to create the function that takes a list of input and converts
-      it to a data struture that is passed to your process/1 function.
+      You'll need to create the function that takes a list of input and preforms the appropriate actions.
       """
 
       IO.warn(message, Macro.Env.stacktrace(env))
-
-      quote do
-        @doc false
-        def init(init_arg) do
-          init_arg
-        end
-
-        defoverridable init: 1
-      end
     end
   end
 end